feat:修改订单字段类型

+++ src/main/java/com/elong/hotel/hotelconfirm/examorder/po/ExamOrderPo.java
@@ -34,7 +34,7 @@ public class ExamOrderPo implements Serializable{
 	/**
 	 * 订单编号
 	 */
-	private Long reserNo;
+	private Integer reserNo;
 	
 	/**
 	 * 订单状态
@@ -270,11 +270,11 @@ public class ExamOrderPo implements Serializable{
 		this.id = id;
 	}
 
-	public Long getReserNo() {
+	public Integer getReserNo() {
 		return reserNo;
 	}
 
-	public void setReserNo(Long reserNo) {
+	public void setReserNo(Integer reserNo) {
 		this.reserNo = reserNo;
 	}
